Changed key type of Index to Int (because it is supposed to map join keys, which are also Int)

+++ src/engine/Code/include/engine/containers/Index.hpp
@@ -7,7 +7,7 @@ namespace containers
 {
 // -------------------------------------------------------------------------
 
-typedef std::unordered_map<size_t, std::vector<size_t>> Index;
+typedef std::unordered_map<Int, std::vector<size_t>> Index;
 
 // -------------------------------------------------------------------------
